Refactor by creating package in data layer
