+++ data/src/main/java/com/pocs/data/mapper/PostDetailMapper.kt
@@ -1,6 +1,6 @@
 package com.pocs.data.mapper
 
-import com.pocs.data.model.PostDetailDto
+import com.pocs.data.model.post.PostDetailDto
 import com.pocs.domain.model.post.PostDetail
 
 fun PostDetailDto.toEntity(id: Int) = PostDetail(
