+++ data/src/main/java/com/pocs/data/mapper/PostWriterMapper.kt
@@ -1,6 +1,6 @@
 package com.pocs.data.mapper
 
-import com.pocs.data.model.PostWriterDto
+import com.pocs.data.model.post.PostWriterDto
 import com.pocs.domain.model.post.PostWriter
 
 fun PostWriterDto.toEntity() = PostWriter(
