+++ data/src/main/java/com/pocs/data/model/post/PostAddBody.kt
@@ -1,4 +1,4 @@
-package com.pocs.data.model
+package com.pocs.data.model.post
 
 data class PostAddBody(
     val title: String,
