+++ data/src/main/java/com/pocs/data/source/PostRemoteDataSource.kt
@@ -1,8 +1,8 @@
 package com.pocs.data.source
 
-import com.pocs.data.model.PostAddBody
-import com.pocs.data.model.PostDetailDto
-import com.pocs.data.model.PostListDto
+import com.pocs.data.model.post.PostAddBody
+import com.pocs.data.model.post.PostDetailDto
+import com.pocs.data.model.post.PostListDto
 import com.pocs.data.model.ResponseBody
 
 interface PostRemoteDataSource {
